requirements.txt

# Core FastAPI and Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Data Models and Validation
pydantic==2.7.1
python-dotenv==1.0.1

# Database
pymongo==4.6.1

# Data Science and ML
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
joblib==1.4.2

# Image Processing
Pillow==10.3.0

# API and HTTP
requests==2.32.3

# Visualization and UI
plotly==5.22.0
streamlit==1.36.0
pydeck==0.8.0
folium==0.17.0
streamlit-folium==0.22.0

# Alerts and Notifications
sendgrid==6.11.0

# Utilities
python-dateutil==2.8.2
psutil==6.1.0

# Additional packages that might be needed
matplotlib==3.8.4
seaborn==0.13.2
altair==5.3.0
watchdog==4.0.1